Split Body into smaller component
