+++ tax-calculator-next/components/Body/FormContext.js
@@ -1,0 +1,9 @@
+import { createContext, useContext } from 'react';
+
+const FormContext = createContext();
+
+export const useFormContext = () => {
+    return useContext(FormContext);
+};
+
+export default FormContext;
